requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cloudinary==1.41.0
cryptography==43.0.0
Django==5.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.7
packaging==24.1
paypalrestsdk==1.13.3
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
